Complete logic for a game of Rock, Paper, Scissors
The logic to play one round of the game was implemented in the `playRound` function, which takes the human's choice and the computer's choice as parameters. The function will determine the winner of the round based on the rules of Rock, Paper, Scissors. This function is wrapped inside the playGame function. The playGame function will call the playRound function for each round of the game and also updates the scores accordingly.